Improve styling and refactor a bit
