+++ client/firebase/firebase-api.js
@@ -1,5 +1,5 @@
 import firebase from "firebase/app";
 
-import clientCredentials from "../../client-cred";
+import clientCredentials from "../client-cred";
 
 export default firebase.apps.length > 0 ? firebase.app() : firebase.initializeApp(clientCredentials);
