+++ client/pages/thread.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-
 import {Button} from "antd";
+
 import MainLayout from "../components/MainLayout";
 import Question from "../components/Question";
 import Reply from "../components/Reply";
